MovieInfo/requirements.txt
pymysql
lxml
scrapy
requests
pyexecjs
